Framework: Removed sites-list use from media validation store.
In order to remove sites list the media validation store interface had to be changed so now the "add" action receives the site object and not the siteId. All the places that use MediaAction add (upload files) had to be changed. Also media/utils had to suffer some changes to not be dependent on JetpackSite prototype functions.